Add leafpic
LeafPic is a fluid, material-designed alternative gallery, it also is ad-free
and open source under GPLv3 license.

it has many options to configure look and feel of app, including color themes.

Active development is going on as per the
https://github.com/HoraApps/LeafPic

but for stability tag#v0.6-beta-1 has been taken into account

Signed-off-by: Satish Patel <tosatishpatel@gmail.com>